fix: Update test imports after project reorganization
- Fixed jest.mock paths to use correct relative paths
- Updated tool imports in tests to use ../../src/tools/
- Fixed akamai-client imports in modular-servers tests
- This resolves test failures after moving tests to __tests__ directory
