+++ __tests__/unit/dns-tools.test.ts
@@ -3,7 +3,7 @@ import { listZones, getZone, createZone, listRecords, upsertRecord, deleteRecord
 import { AkamaiClient } from '../../src/akamai-client';
 
 // Mock the AkamaiClient
-jest.mock('../akamai-client');
+jest.mock('../../src/akamai-client');
 
 describe('DNS Tools', () => {
   let mockClient: jest.Mocked<AkamaiClient>;
