+++ __tests__/unit/hostname-management-advanced.test.ts
@@ -12,7 +12,7 @@ import {
 } from '../../src/tools/hostname-management-advanced';
 
 // Mock the AkamaiClient
-jest.mock('../akamai-client');
+jest.mock('../../src/akamai-client');
 
 describe('Advanced Hostname Management Tools', () => {
   let mockClient: jest.Mocked<AkamaiClient>;
