+++ __tests__/unit/mcp-protocol-compliance.test.ts
@@ -8,7 +8,7 @@ import { jest } from '@jest/globals';
 // Mock modules
 jest.mock('@modelcontextprotocol/sdk/server/index');
 jest.mock('@modelcontextprotocol/sdk/server/stdio');
-jest.mock('../akamai-client');
+jest.mock('../../src/akamai-client');
 
 describe.skip('MCP Protocol Compliance', () => {
   let mockServer: any;
