+++ __tests__/unit/property-tools.test.ts
@@ -3,7 +3,7 @@ import { listGroups } from '../../src/tools/property-tools';
 import { AkamaiClient } from '../../src/akamai-client';
 
 // Mock the AkamaiClient
-jest.mock('../akamai-client');
+jest.mock('../../src/akamai-client');
 
 describe('Property Tools', () => {
   let mockClient: jest.Mocked<AkamaiClient>;
