+++ __tests__/unit/rule-tree-advanced.test.ts
@@ -12,7 +12,7 @@ import {
 import { AkamaiClient } from '../../src/akamai-client';
 
 // Mock AkamaiClient
-jest.mock('../akamai-client');
+jest.mock('../../src/akamai-client');
 
 describe('Advanced Rule Tree Management Tools', () => {
   let mockClient: jest.Mocked<AkamaiClient>;
