+++ scripts/send-mcp-command.sh
